LPS-52679 Move configuration from internal and export the package
