+++ modules/portal/portal-soap-extender/src/com/liferay/portal/soap/extender/configuration/JaxWsApiConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.soap.extender.internal.configuration;
+package com.liferay.portal.soap.extender.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -20,21 +20,14 @@ import aQute.bnd.annotation.metatype.Meta;
  * @author Carlos Sierra Andrés
  */
 @Meta.OCD(
-	factory = true,
-	id = "com.liferay.portal.soap.extender.internal.configuration.SoapExtenderConfiguration"
+	id = "com.liferay.portal.soap.extender.configuration.JaxWsApiConfiguration"
 )
-public interface SoapExtenderConfiguration {
+public interface JaxWsApiConfiguration {
 
-	@Meta.AD(required = false)
-	public String[] contextPaths();
+	@Meta.AD(required = true)
+	public String contextPath();
 
-	@Meta.AD(name = "jax.ws.handler.filters", required = false)
-	public String[] jaxWsHandlerFilterStrings();
-
-	@Meta.AD(name = "jax.ws.service.filters", required = false)
-	public String[] jaxWsServiceFilterStrings();
-
-	@Meta.AD(name = "soap.descriptor.builder", required = false)
-	public String soapDescriptorBuilderFilter();
+	@Meta.AD(deflt = "10000", required = true)
+	public long timeout();
 
 }
