+++ modules/portal/portal-soap-extender/src/com/liferay/portal/soap/extender/internal/SoapExtender.java
@@ -18,7 +18,7 @@ import aQute.bnd.annotation.metatype.Configurable;
 
 import com.liferay.portal.dependency.manager.tccl.TCCLDependencyManager;
 import com.liferay.portal.soap.extender.SoapDescriptorBuilder;
-import com.liferay.portal.soap.extender.internal.configuration.SoapExtenderConfiguration;
+import com.liferay.portal.soap.extender.configuration.SoapExtenderConfiguration;
 
 import java.util.Map;
 
@@ -42,7 +42,7 @@ import org.osgi.service.http.whiteboard.HttpWhiteboardConstants;
  * @author Carlos Sierra Andrés
  */
 @Component(
-	configurationPid = "com.liferay.portal.soap.extender.internal.configuration.SoapExtenderConfiguration",
+	configurationPid = "com.liferay.portal.soap.extender.configuration.SoapExtenderConfiguration",
 	configurationPolicy = ConfigurationPolicy.REQUIRE
 )
 public class SoapExtender {
